+++ ch_926.h
@@ -2,7 +2,7 @@
 #define FOUR_COIN_H
 extern int ch_926_init();
 extern int ch_926_close();
-extern int ch_926_run(int *count);
+extern int ch_926_run(uint64_t *count);
 extern int ch_926_accept_all();
 extern int ch_926_reject_all();
 #endif
